Rework directory restructure
In the future, the `pkg/controller/operatingsystemconfig/generator` package can be deleted entirely
