+++ pkg/controller/operatingsystemconfig/generator/generator.go
