+++ pkg/controller/operatingsystemconfig/generator/generator_suite_test.go
@@ -21,7 +21,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestInternal(t *testing.T) {
+func TestGenerator(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Generator Suite")
+	RunSpecs(t, "Controller OperatingSystemConfig Generator Suite")
 }
