+++ pkg/controller/operatingsystemconfig/generator/templates/cloud-init-ubuntu.template
