+++ pkg/controller/operatingsystemconfig/generator/testfiles/cloud-init
