+++ pkg/controller/operatingsystemconfig/generator/testfiles/cloud-init-containerd-provision
