+++ pkg/controller/operatingsystemconfig/generator/testfiles/cloud-init-with-drop-in
